refactor(execution): move bouncer-related code to its file
