+++ crates/blockifier/src/bouncer.rs
@@ -1,12 +1,13 @@
 use std::collections::{HashMap, HashSet};
 
 use cairo_vm::serde::deserialize_program::BuiltinName;
+use cairo_vm::vm::runners::builtin_runner::HASH_BUILTIN_NAME;
+use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use serde::Deserialize;
 use starknet_api::core::ClassHash;
 
 use crate::blockifier::transaction_executor::{
-    get_casm_hash_calculation_resources, get_particia_update_resources, TransactionExecutorError,
-    TransactionExecutorResult,
+    TransactionExecutorError, TransactionExecutorResult,
 };
 use crate::execution::call_info::ExecutionSummary;
 use crate::fee::gas_usage::get_onchain_data_segment_length;
@@ -273,4 +274,51 @@ impl Bouncer {
     pub fn set_accumulated_weights(&mut self, weights: BouncerWeights) {
         self.accumulated_weights = weights;
     }
+}
+
+/// Returns the estimated VM resources for Casm hash calculation (done by the OS), of the newly
+/// executed classes by the current transaction.
+pub fn get_casm_hash_calculation_resources<S: StateReader>(
+    state: &mut TransactionalState<'_, S>,
+    block_executed_class_hashes: &HashSet<ClassHash>,
+    tx_executed_class_hashes: &HashSet<ClassHash>,
+) -> TransactionExecutorResult<ExecutionResources> {
+    let newly_executed_class_hashes: HashSet<&ClassHash> =
+        tx_executed_class_hashes.difference(block_executed_class_hashes).collect();
+
+    let mut casm_hash_computation_resources = ExecutionResources::default();
+
+    for class_hash in newly_executed_class_hashes {
+        let class = state.get_compiled_contract_class(*class_hash)?;
+        casm_hash_computation_resources += &class.estimate_casm_hash_computation_resources();
+    }
+
+    Ok(casm_hash_computation_resources)
+}
+
+/// Returns the estimated VM resources for Patricia tree updates, or hash invocations
+/// (done by the OS), required by the execution of the current transaction.
+// For each tree: n_visited_leaves * log(n_initialized_leaves)
+// as the height of a Patricia tree with N uniformly distributed leaves is ~log(N),
+// and number of visited leaves includes reads and writes.
+pub fn get_particia_update_resources(
+    block_visited_storage_entries: &HashSet<StorageEntry>,
+    tx_visited_storage_entries: &HashSet<StorageEntry>,
+) -> TransactionExecutorResult<ExecutionResources> {
+    let newly_visited_storage_entries: HashSet<&StorageEntry> =
+        tx_visited_storage_entries.difference(block_visited_storage_entries).collect();
+    let n_newly_visited_leaves = newly_visited_storage_entries.len();
+
+    const TREE_HEIGHT_UPPER_BOUND: usize = 24;
+    let n_updates = n_newly_visited_leaves * TREE_HEIGHT_UPPER_BOUND;
+
+    let patricia_update_resources = ExecutionResources {
+        // TODO(Yoni, 1/5/2024): re-estimate this.
+        n_steps: 32 * n_updates,
+        // For each Patricia update there are two hash calculations.
+        builtin_instance_counter: HashMap::from([(HASH_BUILTIN_NAME.to_string(), 2 * n_updates)]),
+        n_memory_holes: 0,
+    };
+
+    Ok(patricia_update_resources)
 }
